Make lcl_head_time (start-time) an integer
As a float, the conversion to string before for the start-time header
results as a representation in scientific notation. OMSP requires an
integral number of seconds since Epoch there, and the server only parses
the non-decimal part of the time, resulting in start_times in the
database order of magnitude earlier than reality.

As an int, it just works (:

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>